requirements.txt
python-telegram-bot==21.5
supabase==2.5.3
python-dotenv==1.0.1
